+++ catalystwan/endpoints/configuration/policy/list/vpn.py
@@ -7,17 +7,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import VPNList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.vpn import VPNList, VPNListEditPayload, VPNListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class VPNListEditPayload(VPNList, PolicyListId):
-    pass
-
-
-class VPNListInfo(VPNList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyVPNList(APIEndpoints, PolicyListEndpoints):
